debt - extract streams logic from buffer into own generic module
